requirements.txt
streamlit
pandas
plotly
scipy
openpyxl
google-generativeai